refactor: centralize Fingerprint type
