+++ types/fingerprint.ts
@@ -1,0 +1,8 @@
+export type Fingerprint = {
+  escenario: string;
+  epoca: string;
+  protagonista: string;
+  dispositivo: string;
+  tono: string;
+  firstSentence?: string;
+};
